Change gitignore

+++ projects/DigitalVideo/projects/BasicVideo/software/System_wrapper/ps7_cortexa9_0/standalone_ps7_cortexa9_0/bsp/ps7_cortexa9_0/include/xil_io.h
@@ -1,5 +1,5 @@
 /******************************************************************************
-* Copyright (c) 2014 - 2020 Xilinx, Inc.  All rights reserved.
+* Copyright (c) 2014 - 2021 Xilinx, Inc.  All rights reserved.
 * SPDX-License-Identifier: MIT
 ******************************************************************************/
 
@@ -29,6 +29,9 @@
 *                         xil_io.h and made them as static inline
 *       am       10/13/20 Changed the return type of Xil_SecureOut32 function
 *                         from u32 to int
+* 7.50  dp       02/12/21 Fix compilation error in Xil_EndianSwap32() that occur
+*                         when -Werror=conversion compiler flag is enabled
+* 7.5   mus      05/17/21 Update the functions with comments. It fixes CR#1067739.
 *
 * </pre>
 ******************************************************************************/
@@ -165,6 +168,7 @@ static INLINE u64 Xil_In64(UINTPTR Addr)
 ******************************************************************************/
 static INLINE void Xil_Out8(UINTPTR Addr, u8 Value)
 {
+	/* write 8 bit value to specified address */
 	volatile u8 *LocalAddr = (volatile u8 *)Addr;
 	*LocalAddr = Value;
 }
@@ -183,6 +187,7 @@ static INLINE void Xil_Out8(UINTPTR Addr, u8 Value)
 ******************************************************************************/
 static INLINE void Xil_Out16(UINTPTR Addr, u16 Value)
 {
+	/* write 16 bit value to specified address */
 	volatile u16 *LocalAddr = (volatile u16 *)Addr;
 	*LocalAddr = Value;
 }
@@ -202,6 +207,7 @@ static INLINE void Xil_Out16(UINTPTR Addr, u16 Value)
 ******************************************************************************/
 static INLINE void Xil_Out32(UINTPTR Addr, u32 Value)
 {
+	/* write 32 bit value to specified address */
 #ifndef ENABLE_SAFETY
 	volatile u32 *LocalAddr = (volatile u32 *)Addr;
 	*LocalAddr = Value;
@@ -224,6 +230,7 @@ static INLINE void Xil_Out32(UINTPTR Addr, u32 Value)
 ******************************************************************************/
 static INLINE void Xil_Out64(UINTPTR Addr, u64 Value)
 {
+	/* write 64 bit value to specified address */
 	volatile u64 *LocalAddr = (volatile u64 *)Addr;
 	*LocalAddr = Value;
 }
@@ -249,8 +256,10 @@ static INLINE int Xil_SecureOut32(UINTPTR Addr, u32 Value)
 	u32 ReadReg;
 	u32 ReadRegTemp;
 
+	/* writing 32 bit value to specified address */
 	Xil_Out32(Addr, Value);
 
+	/* verify value written to specified address with multiple reads */
 	ReadReg = Xil_In32(Addr);
 	ReadRegTemp = Xil_In32(Addr);
 
@@ -298,8 +307,8 @@ static INLINE __attribute__((always_inline)) u32 Xil_EndianSwap32(u32 Data)
 
 	/* byte swap each of the 16 bit half words */
 
-	LoWord = (((LoWord & 0xFF00U) >> 8U) | ((LoWord & 0x00FFU) << 8U));
-	HiWord = (((HiWord & 0xFF00U) >> 8U) | ((HiWord & 0x00FFU) << 8U));
+	LoWord = (u16)(((LoWord & 0xFF00U) >> 8U) | ((LoWord & 0x00FFU) << 8U));
+	HiWord = (u16)(((HiWord & 0xFF00U) >> 8U) | ((HiWord & 0x00FFU) << 8U));
 
 	/* swap the half words before returning the value */
